python-bot/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.9.0
gunicorn==20.1.0
